requirements.txt
langchain 
langchain-community 
langchain-google-genai  
google-generativeai 
streamlit 
python-dotenv 
PyPDF2 
docx2txt 
faiss-cpu
uvicorn
pydantic 
python-dateutil 
huggingface-hub 
transformers 
fastapi 
sentence-transformers
chromadb
